Extract out `Evaluatable` so that useless fields aren't included.
