+++ src/main/java/com/eliotlash/molang/ast/EvaluatableStmt.java
@@ -1,0 +1,17 @@
+package com.eliotlash.molang.ast;
+
+import java.util.List;
+import java.util.Objects;
+
+public class EvaluatableStmt implements Evaluatable {
+
+    private final List<Stmt> stmts;
+
+    public EvaluatableStmt(List<Stmt> stmts) {
+        this.stmts = Objects.requireNonNull(stmts);
+    }
+
+    public double evaluate(Evaluator evaluator) {
+        return evaluator.evaluate(this.stmts);
+    }
+}
